Aligns reductions with 2023.12 array API spec
Floating point data types are no longer promoted based on item size
